+++ vignettes/getting-started.R
@@ -49,7 +49,7 @@ school_svy <- as_survey_srs(
 school_svy
 
 ## ----calibrated---------------------------------------------------------------
-ns_wave1_svy <- as_survey_calibrated(ns_wave1, weights = weight)
+ns_wave1_svy <- as_survey_nonprob(ns_wave1, weights = weight)
 
 ns_wave1_svy